+++ NEWS.md
@@ -2,6 +2,9 @@
 
 * Set dependency `brickster (>= 0.2.7)`
 * Add github templates for issues, features and PRs
+* Updated `volume_methods` to use zephyr::msg_info(), replacing cli::cli_alert()
+* Updated `volume_methods`, `table_utils` and `table_methods` to use zephyr::get_option() replacing a bool
+* Updated tests to use zephyr-option, with verbosity = quiet
 
 # connector.databricks 0.0.5